+++ src/components/ui/Button.tsx
@@ -1,6 +1,6 @@
 import type { ButtonHTMLAttributes } from 'react';
 
-import { cn } from '../lib/cn';
+import { cn } from '@/lib/cn';
 
 type ButtonVariant = 'v' | 'w' | 'wv' | 'add' | 'icon';
 // v: violet, w: white, wv: bg-white text-violet, add: dashboard add
