+++ src/components/ui/Field.tsx
@@ -2,7 +2,7 @@
 
 import React from 'react';
 
-import { cn } from '../lib/cn';
+import { cn } from '@/lib/cn';
 
 type InjectedProps = {
   id?: string;
