+++ src/components/ui/Input.tsx
@@ -3,9 +3,9 @@
 import type { ComponentPropsWithoutRef } from 'react';
 import React, { useId, useState } from 'react';
 
-import Visibility from '../assets/Visibility.svg';
-import VisibilityOff from '../assets/VisibilityOff.svg';
-import { cn } from '../lib/cn';
+import Visibility from '@/assets/Visibility.svg';
+import VisibilityOff from '@/assets/VisibilityOff.svg';
+import { cn } from '@/lib/cn';
 import Button from './Button';
 import SelectInput from './SelectInput';
