+++ src/lib/cn.ts
@@ -1,5 +1,5 @@
-import { clsx } from "clsx";
-import { twMerge } from "tailwind-merge";
+import { clsx } from 'clsx';
+import { twMerge } from 'tailwind-merge';
 
 export function cn(...inputs: (string | undefined | null | boolean)[]) {
   return twMerge(clsx(inputs));
